share constructor
