+++ pandas/tests/frame/methods/test_astype.py
@@ -762,11 +762,6 @@ class TestAstypeCategorical:
 class IntegerArrayNoCopy(pd.core.arrays.IntegerArray):
     # GH 42501
 
-    @classmethod
-    def _from_sequence(cls, scalars, *, dtype=None, copy=False):
-        values, mask = cls._coerce_to_array(scalars, dtype=dtype, copy=copy)
-        return IntegerArrayNoCopy(values, mask)
-
     def copy(self):
         assert False
